main script and material laws
main3D.m is the main script for homogenization of elasticity at finite strains in 3D. The code is for publication version and thus is written as compact as possible. The two files neo_hookean.m and saint_venant.m declare two functions representing the Neo-Hookean and Saint-Venant material laws. The key input is the deformation gradient and the outputs are the first Piola-Kirchhoff stress and the elastic tangent (Hessian of the energy density).